09.17.2021 Commit
- Fixed loading problem when loading game map and blocks. But there is
still some error when loading a saved game after navigating the menu and
going back and forth more than once. I am assuming it is caused by the
boolean condition within the startLoadGame method. It loads the block
accurately for the first loading but after that it doesn't load the
block correctly.

- Game over condition is added. When any points from the last drop go
above the upper limit of the map size, game is forced to stop.

- Saved files are added as they are created and updated whenever user
decides to save the game. As of now, this game is a single player game
which means there is one file that is being overwritten as the player
saves the game more than once.